Rename ExampleBuildApplicationCommandLineTool to ExampleCommandLineInterfaceTool
