+++ ExamplePlugins/README.md
@@ -3,7 +3,7 @@
 
 This project contains several examples which extend the editor:
 
-1. ExampleBuildApplicationCommandLineTool - Extends labviewnxgcli.exe with two commands - "example-build-application" and "example-build-library". See README.md in the folder for more information.
+1. ExampleCommandLineInterfaceTool - Extends labviewnxgcli.exe with two commands - "example-build-application" and "example-build-library". See README.md in the folder for more information.
 1. ExampleCommandPaneContent - Implements several commands which extend right click menus and the right rail of several nodes
 1. ExampleCustomButtonsToolWindow - Implements a tool window accessible from the Tools Pane. The tool window imports an arbitrary number of buttons via MEF. One example button is provided, which generates DFIR from the open document hierarchy.
 1. ExampleDiagram - A simple Diagram editor without any semantics
